[TRAFODION-1453] Improvements in Hive Scan
The null byte is now skipped over to look for the column
and record delimiter while scanning the hdfs row buffer.
The null byte will be part of the column value.
The behavior of null byte in the column value is unknown.
Skipping over null byte is now in sync with other tool's
behavior.

Also, improved the code to scan the buffer for both record and
column delimiter once. Earlier, the buffer was scanned twice;
once for the record delimiter and another for column delimiter.

No errors are returned when null byte is found.

A sql error 7002 is now reported when a unspported row format type is accessed
from Trafodion.

*** ERROR[7002] The input row format of the hive table is not yet supported.
(cherry picked from commit 5d4e0a927e801d78d97e36938d6bf2261f8e9769)