SPLIT BY clause and growing salted tables (Mantis #15)
Changes for EsgynDB 2.0.

This commit contains two features that are related to how we create
the initial region splits when we create a Trafodion table.

The features are described in TRAFODION-50 and TRAFODION-781, two
JIRAs that stay unresolved, since we implement the features in
EsgynDB.

https://www.esgyn.com/mantis/view.php?id=15

https://issues.apache.org/jira/browse/TRAFODION-50
https://issues.apache.org/jira/browse/TRAFODION-781

TRAFODION-50 requests a way to pre-split a non-salted table
-----------------------------------------------------------

Until now, we only pre-split salted Trafodion tables. To change that,
we need a way to specify the region boundaries in terms of the clustering
key. This is done with the new SPLIT BY clause. The syntax is similar
to RANGE PARTITION BY, except that no location clause is used. See the
JIRA for a more detailed description.

The text of the SPLIT BY clause is stored in the TEXT metadata table,
with a TEXT_TYPE of 6.

My reason for choosing a new syntax is that I would like to reserve
PARTITION BY for something that is tightly managed by Trafodion, so
that it can be used to push operations down to the region servers,
from full GROUP BY to joins on co-located tables.  SPLIT BY is not
tightly managed, the region servers can split regions further.

Like for salted tables, if SPLIT BY is specified, we use the
"constant size" split policy for the new table, to avoid agressive
splitting of these pre-created regions by HBase.

TRAFODION-781 is about future growth of salted tables
-----------------------------------------------------

This is something Rohit pointed out. To allow for future growth, we
should have a larger number of salt values. Initially, there may not
be a need to create a separate HBase region for each one of these. We
could allow the user to specify two numbers at table creation time a)
the number of salt values and b) the number of HBase regions to
create.

New syntax for the SALT clause:

  SALT USING <p> PARTITIONS [ IN <r> REGION[S] ] [ ON (<column(s)>) ]
                            ********************

If the new clause is not specified, the behavior will be as today, the
table is pre-split into p regions. If specified, r must be greater
than or equal to 1 and less than or equal to p. Note that p does not
have to be a multiple of r, although it is certainly a good idea to
choose it that way. If r = 1, we will not apply the
org.apache.hadoop.hbase.regionserver.ConstantSizeRegionSplitPolicy
split policy, like we do for other salted tables.

The existing CQD TRAF_NUM_OF_SALT_PARTNS can be used to salt tables
automatically, without a SALT clause. There is now a new CQD,
TRAF_NUM_OF_SALT_REGIONS, that goes with TRAF_NUM_OF_SALT_PARTNS
and can be used to set the number of regions for such automatically
salted tables.

Like the SPLIT BY clause, the "IN <r> REGIONS" clause is stored
in the TEXT metadata table, using the same TEXT_TYPE 6 as SPLIT BY.

See core/sql/regress/seabase/TEST015 for examples of these new features.