Backup-Restore improvements
1) Backup-Restore (B-R) speed improved by using file copy instead if MapReduce when exporting snapshots of small tables
2) B-R will start Trafodion after all snapshots have been taken, reducing downtime required.

Item 1) is implemented by adding small changes to the existing HBase ExportSnapshot class and renaming the whole class
to be in the Trafodion package. The new java class is packaged in a Trafodion utility jar and added to HBase' classpath.
B-R will call the new class through the hbase binary, as hbase user. HBase classpath has been modified for
install_local_hadoop. Copying this jar to an existing hbase jar directory for cluster installs remains to be done.