Mantis-78 Wrong answers in queries that do range repartitioning.
There were two issues when building the binary start key values for
range partitioning when these values were taken from histogram stats.

First, the constant strings stored in the histogram need to be cast
to the correct column types, since the constants may have a smaller
type (e.g. smallint for value 100 for an integer column or largeint
for the max value 2147483647).

Second, the code used the "+=" operator on an NAString with a
sequence of binary values, which sometimes truncated the values
when they had embedded bytes with the value zero.

Also added some code to make debugging issues with binary key values
a bit easier.