Enable upsert with subquery, when table has index.
upsert on a table with index is transformed to merge. merge does not support
subquery in on, set and insert values clause in general. In this change
we do 2 things
a) make sure that when upsert is transformed to merge there will always
be a using clause. In other words, the row to be upserted will be flowed
to merge operator through a TSJ. This is now true even if there is only
one row to be upserted. Previously the one row case was treated differently.
But that approach cannot handle subqueries in that one row
b) merge restrictions for subquery and udf are relaxed if we know the
merge was created for an upsert. This is valid because an upsert creates a
specific type of merge where the ON, SET and INSERT VALUES clauses use only
vids from the using clause. This is a specific and simpler situation of the
general MERGE statement. With upsert and the idea mentioned in item a) above
we can be sure that subquery will always be evaluated before we get to
the merge statement. Even optimizer will not rearrange the joins because
the merge has specified the subquery as its characteristic input