Add RPM build to the package-all target
This will ensure that the RPM packaging is exercised as part of regular
automated builds. Further infrastructure changes will be needed to use
RPM packages in test process.

The package-all target will continue to generate tar files as before,
since the rpmbuild target uses the tar files to generate RPM.