Fix for merge from trafodion
From Prashanth:

Basically drop schema cascade involves histogram table getting updated and dropped
in the same transaction. With ddl transactions on, there is region observer code
that intercepts region close ( when initiated by hmaster as part of hadmin disable
table or region split) if there is any active transaction in the same region.
Region server invokes the region observer code correctly in case of hbase 1.0 but
not in hbase 0.98. This is why this problem did not show up in trafodion branch but
showed up in enterprise branch.

This patch includes the following:
1. Drop schema cascade drops histograms tables at the end after dropping user tables.
2. Region observer intercept code now checks if a table has IUD operation and same table is getting dropped.
3. AQREntry(   8606,     73,      3,     60,      0,   0, "",    0,     0)
4. Minor updates to regress expected file.

Thanks to Anoop, Narendra, Shang Sheng, Sean , Sandhya for all the help.