Mantis-268 and Mantis-230: Key of Hive tables
Trafodion tables need to have a unique key.

So far, we worked around that for Hive tables by putting all the
columns as key columns (sort columns only for bucketed, sorted
tables). That lead to incorrect results:

Mantis-230 shows this directly, we eliminated a groupby on
the sort columns, which we take as key despite the fact that
they are not unique.

Mantis-268 shows a case where we need a unique key of a table to do
subquery unnesting. It currently asserts because the unique key
includes virtual columns.

Now that we have virtual columns providing unique keys, the fix is to
use two virtual columns as a unique key, this could be either the
standard Hive virtual columns
(INPUT__FILE__NAME, BLOCK__OFFSET__INSIDE__FILE)
or two virtual columns we added in Trafodion:
(INPUT__RANGE__NUMBER, ROW__NUMBER__IN__RANGE).

New CQDs:

HIVE_USE_PERSISTENT_KEY:
   OFF: Use the Trafodion virtual columns, which are much
        shorter, but can vary from one scan to another, so
        they would lead to incorrect results with a self-join
        on these columns.
   ON:  Use Hive virtual columns, which are the same for
        different scans of the same table.

HIVE_USE_SORT_COLS_IN_KEY:
   OFF: Don't add sort columns of bucketed and sorted tables.
   ON:  Do add sort columns of bucketed and sorted tables as
        a key prefix.

I also added a temporary kludge to avoid the error message
seen in Mantis-268, this needs to be removed later. We use
this mostly for TPC benchmarks, all of which have unique rows,
so we should be ok for those.