Mantis-373 crash in generator for nested join w/partitioned Hive table
When we tried to assign the scan info for a partitioned Hive table
(which includes assigning the values of the partition columns), we ran
into a NULL pointer. This is probably a merge conflict. Qifan and I
both made changes to this code and my new constructor accepted a
default argument (NULL for non-partitioned tables), so we didn't find
this conflict at compile time.