PowerPC support: 3rd installment of changes in the Advent2.2 branch.
All the important changes that were made to build on PowerPC are in.
To compile with GCC 4.8 on PowerPC, the switch -Werror was removed
BUT NOT in these checkins. We will need to evaluate (when building
on a PowerPCC machine) whether/where(which component's Makefile)
the switch -Werror is needed.