Error 1034 returned trying view logs as the Admin user
The error "1034 Unable to obtain privileges" is returned when trying to
execute the predefined table mapping function EVENT_LOG_READER.  This udf is a
special predefined table mapping function that is not defined in the metadata.
Therefore, when the privilege checking code is called, there is no metadata
and there is no place to attach privilege descriptors.  The only user that
can use it at this time is DB__ROOT.

The solution is to record these predefined udf's in the metadata. As a
temporary solution, this delivery allows the DB__ADMIN  user in addition to
DB__ROOT to perform the logging requests.  This change allows EsgynDB Manager to
work properly.