Mantis-520 Prevent error 9232 on IUS following failure
Moves the code that restores the row in SB_PERSISTEN_SAMPLES for the source
table of the IUS to ~HSGlobalsClass, so it will be executed even when an
exception or internal failure occurs. Also provides checks to make sure
the row is not changed back to its original state if IUS fails before
modifying it.