Free IUS memory and reset population status between batches
When memory requirements cause incremental update statistics to split
the columns of a table into multiple batches, the 2nd batch fails an
assertion because the in-memory delete and insert tables are already
marked as populated. This change resets the status in selectIUSBatch(),
so each batch can populate its columns. Also, columns processed in a
batch by IUS were not having their memory deallocated after they were
processed by IUS, and this has been fixed as well.