Prevent corruption of ULOG by Incremental Update Statistics
The title of a timing event was being written to a buffer that was
then used for something else before registering the event in the
log. The event showed up in the log with garbage from the
intervening usage. Also, the end of the timing event was missed
due to early exit for an error, resulting in an imbalance of
stop/start events. Both problems are corrected with this fix.