[Mantis-544] Allow sampling pushdown to HBase for single-node queries with predicates
Removes the restriction in RelSample::bindNode() that prevents sampling
in HBase rather than in Trafodion if the scan has predicates. This was
causing IUS queries, which include a Where clause to identify rows
to insert/remove from the persistent sample, to use a sample node
in the query tree instead of the vastly more efficent use of a
RandomRowFilter in HBase. Applying a random sampling operator is
equivalent to adding another conjunct to the scan node's predicates,
and it doesn't matter if it is applied before or after other conjuncts.