MANTIS-0000552 Insufficient privileges on sample table for Update Statistics
User1 created a table
User2 is granted MANAGE_STATISTICS privilege
User2 tried an update statistics on User1's table and got an error:

*** ERROR[9214] Object TRAFODION.xxxxxxxx.SQLMX_8723571504151469986_1467338485_57
6325 could not be created. [2016-07-01 02:01:34]
*** ERROR[1017] You are not authorized to perform this operation.

User2 should be able to update statistics.

When running utility commands like update statistics as a user that does not
have privileges, the utility sets a special parserflag that allows these
operations to proceed even though the requester does not have privs.  We are not
checking this parserflag when we create the sample table and therefore the table
create is failing because the requestor does not have privilege.