Change to support the latest version of HDP 2.3.
The latest version of HDP 2.3 contains Hortonworks' HBase
version: 1.1.2.2.3.6.0-3796 with changes to the HTable ctor
that we were using.

The HTable ctor that we were using was tagged:
InterfaceAudience.Private and so not for 'public' consumption.

The change uses a ctor that is not tagged: InterfaceAudience.Private
but is still 'deprecated'.
(cherry picked from commit e2d19b4f836f81236e7912b7ddab2726f5bd0ce9)