[Mantis-589] Avoid inconsistent IUS memory allocations
Three separate in-memory data sets are used in the execution of
incremental Update Stats. Each pass includes as many columns as are
estimated to fit in memory, but sometimes the estimate is too
liberal and an allocation failure occurs. In the present case, an
allocation for one of the data sets occurred after another had
been done successfully. The recovery code operated independently
on the 3 data sets, resulting in some phases of IUS being
performed on some columns but not others, and leading in this
case to a select statement with an empty SELECT list. This was
all due to an existing allocation function for internal sort
being used, which only dealt with a single data set at a time. Its
independent use on 3 separate but related data sets led to the
inconsistent state. It was replaced with a new allocation
algorithm that coordinates among the 3 data sets and keeps them
consistemt.