Mantis-600, Mantis-560 basic support for NULL Hive partitioning keys
AdvEnt2.1 does not support Hive partitions for NULL values or invalid
values. This fix adds some basic support for NULL values, without
claiming that this is generally supported now. It also does not
support invalid partition values, such as 'abc' for an integer
partition column. Error messages may still be bad, for example asserts
or cores. But this simple fix should give us 80% of the bang for 10%
of the buck.

Basically, NULL values get translated to __HIVE_DEFAULT_PARTITION__
and vice versa when inserting and selecting from partitioned Hive
tables.