[TRAFODION-2129] Trafodion to avoid use of deprecated HBase APIs/Classes
References to the deprecated classes HConnection, HConnectionManager and
HBaseAdmin are reworked to use Connection, ConnectionFactory and Admin classes
respectively.

A connection object is created and it is passed around to ensure that the
same connection object is used within the process. Connection object is thread safe.

Admin object is created and closed after its use immediately.
Creation of admin object is lightweight and the object is not thread safe

Also fixed the random hang at the time of rollback transaction waiting for the response from the TM process.
The TM process hung because it assumed that RS didn't respond back when there is an exception in DDL
rollback or end-point coprocessor call.
(cherry picked from commit 7fe96649f0155bfcc9a6d86328658ac17786258e)